requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.18.0
matplotlib>=3.4.0
nibabel>=3.2.0
scipy>=1.7.0
tqdm>=4.62.0
tensorboard>=2.7.0
Pillow>=8.3.0
SimpleITK>=2.1.0
monai>=0.7.0
einops>=0.4.0
wandb>=0.12.0
pyyaml>=5.4.0
